requirements.txt
exifread
pyexiftool